[FIX] l10n_my_edi: relax credit note check
So far, we have enforced that credit notes must be
linked to an invoice that has been sent to MyInvois.

In real life, this is too strict. Users could be
issuing credit notes for invoices created before
MyInvois was used, in which case these must still
be reported to MyInvois even though the original
invoice isn't on the platform.

This also fixes a small issue with the issue date,
which was set to the invoice date up until now but
on their platform they explicitly ask for today in
UTC timezone.

task-4889022

closes odoo/odoo#216710

X-original-commit: 6bd676d5fddb7f77f2128dc228b1d7501ad9d3ac
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>