[REV] google_calendar: duplicated events
This commit reverts odoo/odoo#208302, odoo/odoo#188848 and odoo/odoo#192876.

Issues have been found on the synchronization with Google related
to the duplication of events for synchronized users. As it is an
urgent matter and these two fixes were merged too close to prior
of the first problem report, we're dropping them for additional
investigation.

task-4873605

...

closes odoo/odoo#215449

Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>