[PERF] account: avoid O(n²) in tax detail with tax groups
When building `base_tax_line_mapping`, the JOIN on
`affecting_base_tax_ids` was effectively scanning the CTE as many times
as there were tax lines in the domain. This lead to computing a number
of rows equivalent to the number of base lines times the number of tax
lines; so O(n²).
Using a LATERAL JOIN helps in avoiding that behavior and only getting
the lines needed in each loop.

X-original-commit: 4bb30d9e258036e98f49af8c10b7c297cae2ac0c
Part-of: odoo/odoo#215992
Related: odoo/enterprise#88593
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>