[FIX] mail: subscribe once to multiple record activities
When creating several activities linked to the same record you may
encouter a crash due to several followers being inserted for the
same record. Indeed everything is done at the same time and all
data is not already flushed.

It is actually better to uniquify list of records to subscribe, so that
subscription is done once / user.

Task-3777606

X-original-commit: odoo/odoo@717ea8db21e31e51e671635ba807ad13a7618ac4
Part-of: odoo/odoo#217688
Related: odoo/enterprise#89613
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>