[FIX] html_editor, project: prevent error when saving after applying history
Problem:
When applying a specific version from history in `task.description`,
then updating and saving it, an error appears stating the content was
saved from a different history model.

Cause:
When inserting versioned content, the required
`data-last-history-steps` attribute is not included.

Solution:
Ensure the latest `data-last-history-steps` is added when restoring a
version.

Steps to reproduce:
- Open Project > any task
- Change the description
- Save
- Open version history and apply any version
- Update the description
- Save
- Error appears

opw-4829553

closes odoo/odoo#216370

Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>